Rename package and clases
