+++ phpspec.yml
@@ -1,5 +1,5 @@
 suites:
   laracasts_suite:
-    namespace: Dinkbit
+    namespace: NotifyMeHQ
     src_path: src
     spec_prefix: spec
