--- src/Dinkbit/Notifyme/Contracts/Gateway.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Dinkbit\Notifyme\Contracts;
-
-interface Gateway
-{
-}
