+++ src/NotifyMeHQ/NotifyMe/Contracts/Gateway.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace NotifyMeHQ\NotifyMe\Contracts;
+
+interface Gateway
+{
+}
