+++ src/NotifyMeHQ/NotifyMe/Gateways/PagerDuty.php
@@ -1,46 +1,32 @@
 <?php
 
-namespace Dinkbit\Notifyme\Gateways;
+namespace NotifyMeHQ\NotifyMe\Gateways;
 
-use Dinkbit\Notifyme\Contracts\Notifier;
-use Dinkbit\Notifyme\Response;
+use NotifyMeHQ\NotifyMe\Contracts\Notifier;
+use NotifyMeHQ\NotifyMe\Response;
 
-class HipChat extends AbstractGateway implements Notifier
+class PagerDuty extends AbstractGateway implements Notifier
 {
     /**
      * Gateway API endpoint.
      *
      * @var string
      */
-    protected $endpoint = 'https://api.hipchat.com';
+    protected $endpoint = 'https://events.pagerduty.com/generic/{version}';
 
     /**
      * Gateway display name.
      *
      * @var string
      */
-    protected $displayName = 'hipchat';
+    protected $displayName = 'pagerduty';
 
     /**
-     * HipChat API version.
+     * Gitter API version.
      *
      * @var string
      */
-    protected $version = 'v2';
-
-    /**
-     * HipChat message background color.
-     *
-     * @var string
-     */
-    protected $colors = [
-        'yellow',
-        'red',
-        'gray',
-        'green',
-        'purple',
-        'random',
-    ];
+    protected $version = '2010-04-15';
 
     /**
      * {@inheritdoc}
@@ -61,11 +47,9 @@ class HipChat extends AbstractGateway implements Notifier
     {
         $params = [];
 
-        $room = $this->array_get($options, 'to', '');
-
         $params = $this->addMessage($message, $params, $options);
 
-        return $this->commit('post', $this->buildUrlFromString("room/{$room}/message"), $params);
+        return $this->commit('post', $this->buildUrlFromString("create_event.json"), $params);
     }
 
     /**
@@ -79,21 +63,13 @@ class HipChat extends AbstractGateway implements Notifier
      */
     protected function addMessage($message, array $params, array $options)
     {
-        $params['auth_token'] = $this->array_get($options, 'token', $this->config['token']);
-
-        $params['id'] = $this->array_get($options, 'to', '');
-        $params['from'] = $this->array_get($options, 'from', $this->config['from']);
-
-        $color = $this->array_get($options, 'color', 'yellow');
-
-        if (! in_array($color, $this->colors)) {
-            $color = 'yellow';
-        }
-
-        $params['color'] = $color;
-        $params['message'] = $message;
-        $params['notify'] = $this->array_get($options, 'notify', false);
-        $params['message_format'] = $this->array_get($options, 'format', 'text');
+        $params['service_key'] = $this->array_get($options, 'token', $this->config['token']);
+        $params['incident_key'] = $this->array_get($options, 'to', 'NotifyMe');
+        $params['event_type'] = $this->array_get($options, 'event_type', 'trigger');
+        $params['client'] = $this->array_get($options, 'client', null);
+        $params['client_url'] = $this->array_get($options, 'client_url', null);
+        $params['details'] = $this->array_get($options, 'details', null);
+        $params['description'] = $message;
 
         return $params;
     }
@@ -105,24 +81,23 @@ class HipChat extends AbstractGateway implements Notifier
     {
         $success = false;
 
-        $token = $params['auth_token'];
-
-        unset($params['auth_token']);
-
         $rawResponse = $this->getHttpClient()->{$method}($url, [
             'exceptions'      => false,
             'timeout'         => '80',
             'connect_timeout' => '30',
             'headers'         => [
                 'Content-Type'  => 'application/json',
-                'Authorization' => 'Bearer '.$token,
             ],
             'json' => $params,
         ]);
 
-        if ($rawResponse->getStatusCode() == 204) {
+        if ($rawResponse->getStatusCode() == 200) {
             $response = [];
             $success = true;
+        } elseif ($rawResponse->getStatusCode() == 404) {
+            $response['error'] = 'Inválid service.';
+        } elseif ($rawResponse->getStatusCode() == 400) {
+            $response['error'] = 'Incorrect request values.';
         } else {
             $response = $this->responseError($rawResponse);
         }
@@ -191,6 +166,6 @@ class HipChat extends AbstractGateway implements Notifier
      */
     protected function getRequestUrl()
     {
-        return $this->endpoint.'/'.$this->version;
+        return str_replace('{version}', $this->version, $this->endpoint);
     }
 }
