+++ src/NotifyMeHQ/NotifyMe/Gateways/Twilio.php
@@ -1,34 +1,39 @@
 <?php
 
-namespace Dinkbit\Notifyme\Gateways;
+namespace NotifyMeHQ\NotifyMe\Gateways;
 
-use Dinkbit\Notifyme\Contracts\Notifier;
-use Dinkbit\Notifyme\Response;
+use NotifyMeHQ\NotifyMe\Contracts\Notifier;
+use NotifyMeHQ\NotifyMe\Response;
 
-class Slack extends AbstractGateway implements Notifier
+class Twilio extends AbstractGateway implements Notifier
 {
     /**
      * Gateway API endpoint.
      *
      * @var string
      */
-    protected $endpoint = 'https://slack.com/api';
+    protected $endpoint = 'https://api.twilio.com';
 
     /**
      * Gateway display name.
      *
      * @var string
      */
-    protected $displayName = 'slack';
+    protected $displayName = 'twilio';
+
+    /**
+     * Twillio API version.
+     *
+     * @var string
+     */
+    protected $version = '2010-04-01';
 
     /**
      * {@inheritdoc}
      */
     public function __construct($config)
     {
-        $this->requires($config, ['token']);
-
-        $config['username'] = $this->array_get($config, 'from', '');
+        $this->requires($config, ['from', 'client', 'token']);
 
         $this->config = $config;
     }
@@ -40,11 +45,15 @@ class Slack extends AbstractGateway implements Notifier
     {
         $params = [];
 
-        $params['unfurl_links'] = true;
+        $this->config['client'] = $this->array_get($options, 'client', $this->config['client']);
+        $this->config['token'] = $this->array_get($options, 'token', $this->config['token']);
+
+        unset($options['client']);
+        unset($options['token']);
 
         $params = $this->addMessage($message, $params, $options);
 
-        return $this->commit('post', $this->buildUrlFromString('chat.postMessage'), $params);
+        return $this->commit('post', $this->buildUrlFromString('Accounts/'.$this->config['client'].'/SMS/Messages.json'), $params);
     }
 
     /**
@@ -58,28 +67,11 @@ class Slack extends AbstractGateway implements Notifier
      */
     protected function addMessage($message, array $params, array $options)
     {
-        $params['token'] = $this->array_get($options, 'token', $this->config['token']);
-        $params['username'] = $this->array_get($options, 'from', $this->config['from']);
-        $params['channel'] = $this->array_get($options, 'to', '');
-        $params['text'] = $this->formatMessage($message);
+        $params['From'] = $this->array_get($options, 'from', $this->config['from']);
+        $params['To'] = $this->array_get($options, 'to', '');
+        $params['Body'] = $message;
 
         return $params;
-    }
-
-    /**
-     * Formats a string for Slack.
-     *
-     * @param string $string
-     *
-     * @return string
-     */
-    public function formatMessage($string)
-    {
-        $string = str_replace('&', '&amp;', $string);
-        $string = str_replace('<', '&lt;', $string);
-        $string = str_replace('>', '&gt;', $string);
-
-        return $string;
     }
 
     /**
@@ -93,12 +85,22 @@ class Slack extends AbstractGateway implements Notifier
             'exceptions'      => false,
             'timeout'         => '80',
             'connect_timeout' => '30',
-            'body'            => $params,
+            'verify'          => true,
+            'auth'            => [
+                $this->config['client'],
+                $this->config['token'],
+            ],
+            'headers' => [
+                'Accept-Charset' => 'utf-8',
+                'Content-Type'   => 'application/x-www-form-urlencoded',
+                'User-Agent'     => 'notifyme/3.12.8 (php '.phpversion().')',
+            ],
+            'body' => $params,
         ]);
 
-        if ($rawResponse->getStatusCode() == 200) {
+        if ($rawResponse->getStatusCode() == 201) {
             $response = $this->parseResponse($rawResponse->getBody());
-            $success = $response['ok'];
+            $success = true;
         } else {
             $response = $this->responseError($rawResponse);
         }
@@ -113,7 +115,7 @@ class Slack extends AbstractGateway implements Notifier
     {
         return (new Response())->setRaw($response)->map([
             'success'       => $success,
-            'message'       => $success ? 'Message sent' : $response['error'],
+            'message'       => $success ? 'Message sent' : $response['message'],
         ]);
     }
 
@@ -154,7 +156,7 @@ class Slack extends AbstractGateway implements Notifier
         $msg .= " (Raw response API {$rawResponse->getBody()})";
 
         return [
-            'error' => $msg,
+            'message' => $msg,
         ];
     }
 
@@ -165,6 +167,6 @@ class Slack extends AbstractGateway implements Notifier
      */
     protected function getRequestUrl()
     {
-        return $this->endpoint;
+        return $this->endpoint.'/'.$this->version;
     }
 }
