+++ src/NotifyMeHQ/NotifyMe/Response.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Dinkbit\Notifyme;
+namespace NotifyMeHQ\NotifyMe;
 
 class Response
 {
